+++ egui/src/data/input.rs
@@ -9,12 +9,12 @@ use crate::math::*;
 /// All coordinates are in points (logical pixels) with origin (0, 0) in the top left corner.
 #[derive(Clone, Debug)]
 pub struct RawInput {
-    /// Is the button currently down?
+    /// Is the mouse button currently down (or a finger, on touch screens??
     /// NOTE: egui currently only supports the primary mouse button.
-    pub mouse_down: bool,
+    pub pointer_button_down: bool,
 
-    /// Current position of the mouse in points.
-    pub mouse_pos: Option<Pos2>,
+    /// Current position of the mouse/touch in points.
+    pub pointer_pos: Option<Pos2>,
 
     /// How many points (logical pixels) the user scrolled
     pub scroll_delta: Vec2,
@@ -56,8 +56,8 @@ impl Default for RawInput {
     fn default() -> Self {
         #![allow(deprecated)] // for screen_size
         Self {
-            mouse_down: false,
-            mouse_pos: None,
+            pointer_button_down: false,
+            pointer_pos: None,
             scroll_delta: Vec2::zero(),
             screen_size: Default::default(),
             screen_rect: None,
@@ -75,8 +75,8 @@ impl RawInput {
     pub fn take(&mut self) -> RawInput {
         #![allow(deprecated)] // for screen_size
         RawInput {
-            mouse_down: self.mouse_down,
-            mouse_pos: self.mouse_pos,
+            pointer_button_down: self.pointer_button_down,
+            pointer_pos: self.pointer_pos,
             scroll_delta: std::mem::take(&mut self.scroll_delta),
             screen_size: self.screen_size,
             screen_rect: self.screen_rect,
@@ -208,8 +208,8 @@ impl RawInput {
     pub fn ui(&self, ui: &mut crate::Ui) {
         #![allow(deprecated)] // for screen_size
         let Self {
-            mouse_down,
-            mouse_pos,
+            pointer_button_down,
+            pointer_pos,
             scroll_delta,
             screen_size: _,
             screen_rect,
@@ -220,10 +220,8 @@ impl RawInput {
             events,
         } = self;
 
-        // TODO: simpler way to show values, e.g. `ui.value("Mouse Pos:", self.mouse_pos);
-        // TODO: `ui.style_mut().text_style = TextStyle::Monospace`;
-        ui.label(format!("mouse_down: {}", mouse_down));
-        ui.label(format!("mouse_pos: {:.1?}", mouse_pos));
+        ui.label(format!("pointer_button_down: {}", pointer_button_down));
+        ui.label(format!("pointer_pos: {:.1?}", pointer_pos));
         ui.label(format!("scroll_delta: {:?} points", scroll_delta));
         ui.label(format!("screen_rect: {:?} points", screen_rect));
         ui.label(format!("pixels_per_point: {:?}", pixels_per_point))