+++ egui/src/memory.rs
@@ -125,7 +125,8 @@ pub(crate) struct Interaction {
 }
 
 impl Interaction {
-    pub fn is_using_mouse(&self) -> bool {
+    /// Are we currently clicking or dragging an egui widget?
+    pub fn is_using_pointer(&self) -> bool {
         self.click_id.is_some() || self.drag_id.is_some()
     }
 
@@ -138,12 +139,12 @@ impl Interaction {
         self.click_interest = false;
         self.drag_interest = false;
 
-        if !prev_input.mouse.could_be_click {
+        if !prev_input.pointer.could_be_click {
             self.click_id = None;
         }
 
-        if !prev_input.mouse.down || prev_input.mouse.pos.is_none() {
-            // mouse was not down last frame
+        if !prev_input.pointer.down || prev_input.pointer.pos.is_none() {
+            // pointer button was not down last frame
             self.click_id = None;
             self.drag_id = None;
         }
@@ -175,7 +176,7 @@ impl Memory {
     ) {
         self.interaction.begin_frame(prev_input, new_input);
 
-        if !prev_input.mouse.down {
+        if !prev_input.pointer.down {
             self.window_interaction = None;
         }
     }