+++ egui/src/widgets/text_edit.rs
@@ -318,45 +318,45 @@ impl<'t> TextEdit<'t> {
         }
 
         if enabled {
-            if let Some(mouse_pos) = ui.input().mouse.pos {
+            if let Some(pointer_pos) = ui.input().pointer.pos {
                 // TODO: triple-click to select whole paragraph
                 // TODO: drag selected text to either move or clone (ctrl on windows, alt on mac)
 
-                let cursor_at_mouse = galley.cursor_from_pos(mouse_pos - response.rect.min);
+                let cursor_at_pointer = galley.cursor_from_pos(pointer_pos - response.rect.min);
 
-                if response.hovered && ui.input().mouse.is_moving() {
+                if response.hovered && ui.input().pointer.is_moving() {
                     // preview:
-                    paint_cursor_end(ui, response.rect.min, &galley, &cursor_at_mouse);
+                    paint_cursor_end(ui, response.rect.min, &galley, &cursor_at_pointer);
                 }
 
                 if response.hovered && response.double_clicked {
                     // Select word:
-                    let center = cursor_at_mouse;
+                    let center = cursor_at_pointer;
                     let ccursorp = select_word_at(text, center.ccursor);
                     state.cursorp = Some(CursorPair {
                         primary: galley.from_ccursor(ccursorp.primary),
                         secondary: galley.from_ccursor(ccursorp.secondary),
                     });
-                } else if response.hovered && ui.input().mouse.pressed {
+                } else if response.hovered && ui.input().pointer.pressed {
                     ui.memory().request_kb_focus(id);
                     if ui.input().modifiers.shift {
                         if let Some(cursorp) = &mut state.cursorp {
-                            cursorp.primary = cursor_at_mouse;
+                            cursorp.primary = cursor_at_pointer;
                         } else {
-                            state.cursorp = Some(CursorPair::one(cursor_at_mouse));
+                            state.cursorp = Some(CursorPair::one(cursor_at_pointer));
                         }
                     } else {
-                        state.cursorp = Some(CursorPair::one(cursor_at_mouse));
+                        state.cursorp = Some(CursorPair::one(cursor_at_pointer));
                     }
-                } else if ui.input().mouse.down && response.active {
+                } else if ui.input().pointer.down && response.active {
                     if let Some(cursorp) = &mut state.cursorp {
-                        cursorp.primary = cursor_at_mouse;
+                        cursorp.primary = cursor_at_pointer;
                     }
                 }
             }
         }
 
-        if ui.input().mouse.pressed && !response.hovered {
+        if ui.input().pointer.pressed && !response.hovered {
             // User clicked somewhere else
             ui.memory().surrender_kb_focus(id);
         }